requirements.txt
openai>=1.0.0
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0

# Optional providers (install as needed):
# anthropic>=0.25.0
